% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Detection.R
\name{cutoff.gen.rec}
\alias{cutoff.gen.rec}
\title{Cutoff Initialization (BH based)}
\usage{
cutoff.gen.rec(Tm, Ta, Tm.star = Inf, Ta.star = Inf)
}
\arguments{
\item{Tm}{a numerical vector representing the auxiliary statistics.}

\item{Ta}{a numerical vector representing the primary statistics.}

\item{Tm.star}{a numerical value representing the threshold found by one dimensional detection based on Tm.}

\item{Ta.star}{a numerical value representing the threshold found by one dimensional detection based on Ta.}
}
\value{
list;
cutoff is a matrix with two columns representing the candidate cutoff set and it should be noted that the cutoffs are ordered for searching convenious;
ind.Tm is an integer vector representing the indexes where the first coordinate of cutoffs, Tm, changes.
}
\description{
Initialize the candidate cutoff set based on primary and auxiliary statistics for searching the optimal threshold.
}
\examples{
library(ggplot2)
set.seed(1234)
Tm <- rnorm(20)
Ta1 <- rnorm(20)
Ta2 <- rnorm(20,Tm,0.5) # depends on Tm
cutoff1 <- cutoff.gen.rec(Tm,Ta1)
cutoff2 <- cutoff.gen.rec(Tm,Ta2)
# Visualize finding cutoffs
Type <- c(rep("Cutoff",dim(cutoff1$cutoff)[1]),rep("Data",20))
qplot(x=c(cutoff1$cutoff[,1],Tm),y=c(cutoff1$cutoff[,2],Ta1),
      colour = Type, shape=Type, xlab="Tm", ylab="Ta")+
      scale_shape_manual(values = c(20,4))
dim(cutoff1$cutoff)[1] # The number of cutoffs
Type <- c(rep("Cutoff",dim(cutoff2$cutoff)[1]),rep("Data",20))
qplot(x=c(cutoff2$cutoff[,1],Tm),y=c(cutoff2$cutoff[,2],Ta2),
      colour = Type, shape=Type, xlab="Tm", ylab="Ta")+
      scale_shape_manual(values = c(20,4))
dim(cutoff2$cutoff)[1] # The number of cutoffs
}
