% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic.R
\name{fdp}
\alias{fdp}
\title{FDP evaluation}
\usage{
fdp(selected, mu)
}
\arguments{
\item{selected}{a vector consists of the indexes of selected signals.}

\item{mu}{a vector of the true population mean.}
}
\value{
FDP: a numerical value, the number of false discoveries/the number of discoveries.
}
\description{
Function to evaluate FDP for upper tail hypothesis.
}
\examples{
mu <- rnorm(100);
selected <- sample(100,10);
fdp(selected,mu)
}
