% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_Setting.R
\name{Init_Setting_1D}
\alias{Init_Setting_1D}
\title{One dimension setting}
\usage{
Init_Setting_1D(
  mu_type = "Medium",
  Cov_type = "Medium",
  mu_gen_machine = "uc.spline",
  magnitude = 1,
  point = point,
  single_bump_prop = 1/20,
  Dist.p = NULL
)
}
\arguments{
\item{mu_type}{a character indicating the type of mu, taking one of the value Sparse, Medium and Dense.}

\item{Cov_type}{a character indicating the type of covariance, taking one of the value Weak, Medium and Strong.}

\item{mu_gen_machine}{a character indicating generating mechanism, taking one of the value uc.spline, uc.unif and mvnorm.}

\item{magnitude}{a numeric value indicating the magnitude of signals.}

\item{point}{an m*1 matrix or numeric vector indicating the point location on one dimensional space.}

\item{single_bump_prop}{a numeric value indicating how many space one bump occupies.}

\item{Dist.p}{a m*m matrix representing the distance between points.}
}
\value{
list; mu is a numeric vector representing the true population mean; Sigma.eps.p is a matrix representing the covariance among noises.
}
\description{
Generate mu & covariance matrix (One dimensional case)
}
\examples{
library(dplyr)
library(fda)
m <- 900 # point size
point <- matrix(seq(0,1,length.out=m), ncol=1)
res <- Init_Setting_1D(mu_type = "Medium",Cov_type = "Medium",
                       mu_gen_machine="uc.spline",point=point)
}
