% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_Setting.R
\name{Init_Setting_2D}
\alias{Init_Setting_2D}
\title{Two dimension setting}
\usage{
Init_Setting_2D(
  mu_type = "Medium",
  Cov_type = "Medium",
  mu_gen_machine = NULL,
  magnitude = 1,
  point = point,
  Dist.p = NULL
)
}
\arguments{
\item{mu_type}{a character indicating the type of mu, taking one of the value Sparse, Medium and Dense.}

\item{Cov_type}{a character indicating the type of covariance, taking one of the value Weak, Medium and Strong.}

\item{mu_gen_machine}{a character indicating generating mechanism, taking one of the value uc.square, uc.circle and uc.mix.}

\item{magnitude}{a numeric value indicating the magnitude of signals.}

\item{point}{an m*2 matrix or numeric vector indicating the point location on two dimensional space.}

\item{Dist.p}{The distance matrix between points.}
}
\value{
list; mu is a matrix representing the true population mean; Sigma.eps.p is a matrix representing the covariance among noises.
}
\description{
Generate mu & covariance matrix (Two dimensional case)
}
\examples{
library(dplyr)
library(fda)
m <- 900 # point size
m_x <- 30
m_y <- 30
point_x <- seq(0,1,length.out=m_x)
point_y <- seq(0,1,length.out=m_x)
point <- cbind(rep(point_x,times=length(point_y)),
               rep(point_y,each=length(point_x)))
res <- Init_Setting_2D(mu_type = "Medium",Cov_type = "Medium",
                       mu_gen_machine="uc.square",point=point)
}
