% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic.R
\name{K.mu.d}
\alias{K.mu.d}
\title{Covariance Matrix for mu}
\usage{
K.mu.d(d, k, rho.mu, sig2.mu)
}
\arguments{
\item{d}{an m*m matrix representing the distance, m is the number of locations.}

\item{k}{a numerical value acting as the power term inside exponential.}

\item{rho.mu}{a numerical value representing the spatial range of mu.}

\item{sig2.mu}{a numerical value representing error variance.}
}
\value{
an m*m matrix representing the covariance for mu.
}
\description{
Function to calculate Covariance matrix for mu (Multivariate normal distribution).
}
\examples{
A <- matrix(runif(4),2)
K.mu.d(A,1,1,1)
}
