% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Detection.R
\name{L.cal}
\alias{L.cal}
\title{False discovery probability}
\usage{
L.cal(tm, ta, mm, normalized.prob, vVa, vVmVa.cov)
}
\arguments{
\item{tm}{a numerical value representing the threshold for the primary statistics.}

\item{ta}{a numerical value representing the threshold for the auxiliary statistics}

\item{mm}{a numerical vector representing the estimated support of the prior distribution for the auxiliary statistics.}

\item{normalized.prob}{a numerical vector representing the corresponding estimated probability of the prior distribution for the auxiliary statistics.}

\item{vVa}{a numerical value representing the variance of the auxiliary statistic. (Usually be one, so we omit using it here.)}

\item{vVmVa.cov}{a numerical value representing the covariance between the primary and auxiliary statistic. (In most case, it is the correlation between them.)}
}
\value{
a numeric value representing the estimated false discovery probability.
}
\description{
Estimated false discovery probability for one location.
}
\examples{
m <- 900
set.seed(1234)
Tm <- rnorm(m)
Ta <- rnorm(m)
mm <- REBayes::GLmix(x = Ta)
normalized.prob <- mm$y / sum(mm$y)
tm0 <- 1
ta0 <- 1
L.val <- L.cal(tm0,ta0,mm,normalized.prob,1,0)
# estimated fdp
(L.val+0.1)/sum(Tm>=tm0&Ta>=ta0)
}
