% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_Setting.R
\name{mu.fix.gen.1D}
\alias{mu.fix.gen.1D}
\title{One dimension setting}
\usage{
mu.fix.gen.1D(
  point,
  method,
  magnitude,
  mu.mean,
  Sigma.mu,
  num_bump = 2,
  single_bump_prop = 1/6
)
}
\arguments{
\item{point}{an m*1 matrix or numeric vector indicating the point location on one dimensional space, m is the number of locations.}

\item{method}{a character indicating the generating mechanism of mu, taking one of uc.spline, uc.unif and mvnorm.}

\item{magnitude}{a numeric value indicating the magnitude of signals.}

\item{mu.mean}{a numeric value shifting the signal (mvnorm mechanism).}

\item{Sigma.mu}{a matrix representing the covariance matrix for mu (mvnorm mechanism).}

\item{num_bump}{a integer indicating the number of bumps.(uc.unif & uc.spline mechanism).}

\item{single_bump_prop}{a numeric value indicating how many space one bump occupies.}
}
\value{
a list; mu is a numeric vector representing the true population mean; pis is a numeric vector representing the probability of being null.
}
\description{
Basic function for initializing signal mu (Fix magnitude): 1 Dimension
}
\examples{
library(dplyr)
library(fda)
point <- sort(runif(100))
res1 <- mu.fix.gen.1D(point = point, method = "uc.spline",
                      magnitude = 1)
res2 <- mu.fix.gen.1D(point = point, method = "uc.unif",
                      magnitude = 1)
Dist.p <- as.matrix(dist(point))
Sigma.mu.p <- K.mu.d(Dist.p, k =1, rho.mu = 0.3, sig2.mu=1)
res3 <- mu.fix.gen.1D(point = point, method = "mvnorm",
                      magnitude = 1, mu.mean=0,
                      Sigma.mu = Sigma.mu.p)
}
