% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_Setting.R
\name{mu.fix.gen.2D}
\alias{mu.fix.gen.2D}
\title{Two dimension setting}
\usage{
mu.fix.gen.2D(point, method, magnitude, mu.mean, Sigma.mu)
}
\arguments{
\item{point}{an m*2 matrix or numeric vector indicating the point location on one, m is the number of locations.}

\item{method}{a character indicating the generating mechanism of mu, taking one of uc.square, uc.circle and uc.mix.}

\item{magnitude}{a numeric value indicating the magnitude of signals.}

\item{mu.mean}{a numeric value shifting the signal (mvnorm mechanism).}

\item{Sigma.mu}{a matrix representing the covariance matrix for mu (mvnorm mechanism).}
}
\value{
a list; mu is a matrix representing the true population mean; pis is a matrix representing the probability of being null.
}
\description{
Init mu (Fix magnitude): 2 Dimension
}
\examples{
library(dplyr)
library(fda)
m <- 900 # point size
m_x <- 30
m_y <- 30
point_x <- seq(0,1,length.out=m_x)
point_y <- seq(0,1,length.out=m_x)
point <- cbind(rep(point_x,times=length(point_y)),
               rep(point_y,each=length(point_x)))
res1 <- mu.fix.gen.2D(point,method="uc.circle",magnitude=1)
res2 <- mu.fix.gen.2D(point,method="uc.square",magnitude=1)
res3 <- mu.fix.gen.2D(point,method="uc.mix",magnitude=1)
}
