% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Detection.R
\name{OneD_Detect}
\alias{OneD_Detect}
\title{One dimension detection for primary statistics}
\usage{
OneD_Detect(Tm, q = 0.1, pis = NULL, const = 0, tau.tm = 1)
}
\arguments{
\item{Tm}{a numerical vector representing the auxiliary statistics.}

\item{q}{a numerical value representing the target fdr level.}

\item{pis}{a numerical vector representing the probability of being null.}

\item{const}{a numerical value ensuring the selection stability, usually be target fdr level.}

\item{tau.tm}{a numerical value representing the censor level for p-values.}
}
\value{
list;
selected is an integer vector consisting of selected locations;
tm.min is the one dimensional threshold;
max.rej is an interger indicating the number of selected locations.
}
\description{
The degenerated 2d-SMT by setting the threshold for Ta to negative infinity.
This is an equivalent expression for BH procedure if we set const=0, tau.tm=1, and pis = 1.
}
\examples{
fdr.level <- 0.1
Tm <- rnorm(100)
res.1D.Tm <- OneD_Detect(Tm,q=fdr.level)
}
