% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Detection.R
\name{OneD_Detect_w}
\alias{OneD_Detect_w}
\title{One dimension detection for primary statistics (weighted BH based)}
\usage{
OneD_Detect_w(
  Tm,
  q = 0.1,
  pis = NULL,
  const = 0,
  ws.fun = NULL,
  ws = NULL,
  tol = 1e-09,
  tau.tm = 1
)
}
\arguments{
\item{Tm}{a numerical vector representing the auxiliary statistics.}

\item{q}{a numerical value representing the target fdr level.}

\item{pis}{a numerical vector representing the probability of being null.}

\item{const}{a numerical value ensuring the selection stability, usually be target fdr level.}

\item{ws.fun}{a function for weighting, depending on the null proportion.}

\item{ws}{a numerical vector indicating the non-null weights.}

\item{tol}{a numerical value to control the calculation precision.}

\item{tau.tm}{a numerical value representing the censor level for p-values.}
}
\value{
list;
selected is an integer vector consisting of selected locations;
tm.min is the one dimensional threshold;
max.rej is an interger indicating the number of selected locations.
}
\description{
This is an equivalent expression for weighted BH procedure if we set const=0 and tau.tm=1.
The weights and pis can be assigned separately.
}
\examples{
fdr.level <- 0.1
Tm <- rnorm(100)
res.1D.w.Tm <- OneD_Detect_w(Tm,q=fdr.level)
}
