% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic.R
\name{Pow}
\alias{Pow}
\title{POWER evaluation}
\usage{
Pow(selected, mu)
}
\arguments{
\item{selected}{a vector consists of the indexes of selected signals.}

\item{mu}{a vector of the true population mean.}
}
\value{
POWER: a numeric value, the number of true discoveries/the number of true signals.
}
\description{
Function to evaluate POWER for upper tail hypothesis.
}
\examples{
 mu <- rnorm(100);
 selected <- sample(100,10);
 Pow(selected,mu)
}
