% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Detection.R
\name{OneD_Detect_Ta}
\alias{OneD_Detect_Ta}
\title{One dimension detection for auxiliary statistics (BH based)}
\usage{
OneD_Detect_Ta(Ta, q = 0.1, ind, pis = NULL)
}
\arguments{
\item{Ta}{a numerical vector representing the auxiliary statistics.}

\item{q}{a numerical value representing the target fdr level.}

\item{ind}{an integer vector representing the indexes for implementing nonparametric empirical Bayes.}

\item{pis}{a numerical vector representing the probability of being null.}
}
\value{
list;
selected is an integer vector consisting of selected locations;
ta.min is the one dimensional threshold for auxiliary statistics;
max.rej is an interger indicating the number of selected locations.
}
\description{
One Dimension Detection based on the auxiliary statistics.
}
\examples{
fdr.level <- 0.1
Ta <- rnorm(100)
res.1D.Ta <- OneD_Detect_Ta(Ta,q=fdr.level,ind=1:100)
}
